Move distribution spot resolvers into dedicated modules
